backend/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
psycopg2-binary==2.9.10
django-cors-headers==4.7.0
pandas==2.2.3
prometheus-client==0.22.0
django-prometheus==2.3.1
gunicorn==23.0.0